made debug_tools a folder to make room for assert_evaluator feature
